ref: Updating tests as part of splitting event_search up
