+++ tests/sentry/api/test_issue_search.py
@@ -1,7 +1,6 @@
 from sentry.api.event_search import (
     AggregateFilter,
     AggregateKey,
-    InvalidSearchQuery,
     SearchFilter,
     SearchKey,
     SearchValue,
@@ -14,6 +13,7 @@ from sentry.api.issue_search import (
     parse_search_query,
     value_converters,
 )
+from sentry.exceptions import InvalidSearchQuery
 from sentry.models.group import STATUS_QUERY_CHOICES
 from sentry.testutils import TestCase
