+++ tests/snuba/api/endpoints/test_organization_events_trends.py
@@ -3,7 +3,7 @@ from datetime import timedelta
 from django.urls import reverse
 
 from sentry.api.endpoints.organization_events_trends import OrganizationEventsTrendsEndpointBase
-from sentry.api.event_search import get_filter
+from sentry.search.events.filter import get_filter
 from sentry.testutils import APITestCase, SnubaTestCase
 from sentry.testutils.cases import TestCase
 from sentry.testutils.helpers import parse_link_header
